+++ apps/oracle-node/oracle-node/src/main/java/bisq/oracle_node/bisq1_bridge/Bisq1BridgeService.java
@@ -18,12 +18,12 @@
 package bisq.oracle_node.bisq1_bridge;
 
 import bisq.bonded_roles.BondedRoleType;
-import bisq.bonded_roles.alert.AlertType;
-import bisq.bonded_roles.alert.AuthorizedAlertData;
 import bisq.bonded_roles.bonded_role.AuthorizedBondedRole;
 import bisq.bonded_roles.bonded_role.AuthorizedBondedRolesService;
 import bisq.bonded_roles.oracle.AuthorizedOracleNode;
 import bisq.bonded_roles.registration.BondedRoleRegistrationRequest;
+import bisq.bonded_roles.security_manager.AlertType;
+import bisq.bonded_roles.security_manager.AuthorizedAlertData;
 import bisq.common.application.Service;
 import bisq.common.encoding.Hex;
 import bisq.common.timer.Scheduler;
