+++ support/src/main/java/bisq/support/security_manager/SecurityManagerService.java
@@ -18,10 +18,10 @@
 package bisq.support.security_manager;
 
 import bisq.bonded_roles.BondedRolesService;
-import bisq.bonded_roles.alert.AlertType;
-import bisq.bonded_roles.alert.AuthorizedAlertData;
 import bisq.bonded_roles.bonded_role.AuthorizedBondedRole;
 import bisq.bonded_roles.bonded_role.AuthorizedBondedRolesService;
+import bisq.bonded_roles.security_manager.AlertType;
+import bisq.bonded_roles.security_manager.AuthorizedAlertData;
 import bisq.common.application.Service;
 import bisq.common.observable.Observable;
 import bisq.common.util.StringUtils;
